Improvements and bugfixes
1) The cbssh.sh utility, found on cbtool/util/, can be used to obtain an ssh prompt on
a VM directly. For instance "~/cbtool/util/cbssh.sh vm_3" will automatically discover the
VM's IP address, login, and (if needed) the name of the Jumphost used to access the VM,
and issue the command accordingly.

2) Improvements on the "typeshow" command. Now the output of this command is divided in
four sections: brief description, general mandatory attributes, AI-specific mandatory
attributes, AI-specific optional attributes.

3) In an attempt to solve an ongoing problem with OpenStack clouds under stress (i.e.,
with hundreds or thousands of VM being deployed), CB will be more tolerant with errors
during "get" or "list" API calls. What seems to be happening is, during a VM deployment,
when CB is polling the Cloud in order to get the instance state, sometimes the "get" (or
"list" API calls return an error). Most of these errors seem to be transient, and CB will
now use a per-VM "get"/"list" error counter (default value is 10) before deciding to
consider a VM failed, and starting a rollback of the Virtual-Application deployment.

4) Several improvements for Cassandra YCSB scripts. First of all, the Virtual Application
deployment is considered complete only after *all* nodes are joined to the cluster (the
utility "nodetool" is employed to this end). Second, we now have a truly idempotent
"restart node" and "restart seed" scripts. At the end of load id execution, these scripts
are re-run, and seeds and nodes are restarted accordingly. The idea here is to prevent
problems stemming from nodes or seeds suddenly dying in the middle of an experiment.

+++ scripts/cassandra_ycsb/cb_restart_node.sh
@@ -30,20 +30,24 @@ sudo mkdir -p ${CASSANDRA_DATA_DIR}/cassandra/commitlog
 sudo mkdir -p ${CASSANDRA_DATA_DIR}/cassandra/saved_caches
 sudo chown -R cassandra:cassandra ${CASSANDRA_DATA_DIR}
 
-sudo service cassandra stop 
-
-sudo mkdir -p ${CASSANDRA_DATA_DIR}
-
 VOLUME=$(get_attached_volumes)
 if [[ $VOLUME != "NONE" ]]
 then
-    if [[ $(check_filesystem $VOLUME) == "none" ]]
-    then
-        syslog_netcat "Creating $CASSANDRA_DATA_FSTYP filesystem on volume $VOLUME"
-        sudo mkfs.$CASSANDRA_DATA_FSTYP $VOLUME
-    fi
-    syslog_netcat "Making $FSTYP filesystem on volume $VOLUME accessible through the mountpoint ${CASSANDRA_DATA_DIR}"
-    sudo mount $VOLUME ${CASSANDRA_DATA_DIR}
+	if [[ $(sudo mount | grep $VOLUME | grep -c $CASSANDRA_DATA_DIR) -eq 0 ]]
+	then
+		sudo service cassandra stop
+		
+		sudo mkdir -p ${CASSANDRA_DATA_DIR}
+				
+	    if [[ $(check_filesystem $VOLUME) == "none" ]]
+	    then
+	        syslog_netcat "Creating $CASSANDRA_DATA_FSTYP filesystem on volume $VOLUME"
+	        sudo mkfs.$CASSANDRA_DATA_FSTYP $VOLUME
+	    fi
+	    
+	    syslog_netcat "Making $FSTYP filesystem on volume $VOLUME accessible through the mountpoint ${CASSANDRA_DATA_DIR}"
+	    sudo mount $VOLUME ${CASSANDRA_DATA_DIR}
+	fi
 fi
 
 pos=1
@@ -66,7 +70,10 @@ done
 #
 # Cassandra will not properly start if the hostname is not in DNS or /etc/hosts
 #
-sudo sh -c "echo ${MY_IP} ${SHORT_HOSTNAME} >> /etc/hosts"
+if [[ $(sudo cat /etc/hosts | grep $MY_IP | grep -c ${SHORT_HOSTNAME}) -eq 0 ]]
+then
+    sudo sh -c "echo ${MY_IP} ${SHORT_HOSTNAME} >> /etc/hosts"
+fi
 
 CASSANDRA_REPLICATION_FACTOR=$(get_my_ai_attribute_with_default replication_factor 4)
 sudo sed -i "s/REPLF/${CASSANDRA_REPLICATION_FACTOR}/g" create_keyspace.cassandra
@@ -75,7 +82,7 @@ CASSANDRA_CONF_PATH=$(get_my_ai_attribute_with_default cassandra_conf_path /etc/
 
 if [[ ! -f $CASSANDRA_CONF_PATH ]]
 then
-	CASSANDRA_CONF_PATH=$(sudo find /etc -name cassandra.yaml)
+    CASSANDRA_CONF_PATH=$(sudo find /etc -name cassandra.yaml)
 fi
 
 #
@@ -85,7 +92,7 @@ sudo sed -i "s/initial_token:$/initial_token: ${my_token//[[:blank:]]/}/g" ${CAS
 sudo sed -i "s/- seeds:.*$/- seeds: $seed_ips_csv/g" ${CASSANDRA_CONF_PATH}
 sudo sed -i "s/listen_address:.*$/listen_address: ${MY_IP}/g" ${CASSANDRA_CONF_PATH}
 sudo sed -i "s/rpc_address:.*$/rpc_address: ${MY_IP}/g" ${CASSANDRA_CONF_PATH}
-sudo sed -i "s/partitioner:.*$/partitioner: org.apache.cassandra.dht.RandomPartitioner/g" ${CASSANDRA_CONF_PATH}	
+sudo sed -i "s/partitioner:.*$/partitioner: org.apache.cassandra.dht.RandomPartitioner/g" ${CASSANDRA_CONF_PATH}    
 TEMP_CASSANDRA_DATA_DIR=$(echo ${CASSANDRA_DATA_DIR} | sed 's/\//_+-_-+/g')
 sudo sed -i "s/\/var\/lib\//${TEMP_CASSANDRA_DATA_DIR}\//g" ${CASSANDRA_CONF_PATH}
 sudo sed -i "s/_+-_-+/\//g" ${CASSANDRA_CONF_PATH}
@@ -101,8 +108,50 @@ sudo rm -rf ${CASSANDRA_DATA_DIR}/cassandra/commitlog/*
 #
 # Start the database
 #
-syslog_netcat "Starting cassandra on ${SHORT_HOSTNAME}" 
-sudo service cassandra start 
+
+FIRST_SEED=$(echo $seed_ips_csv | cut -d ',' -f 1)
+
+wait_until_port_open ${FIRST_SEED} 9160 1 1
+
+STATUS=$?
+    
+if [[ ${STATUS} -eq 0 ]]
+then
+    THRIFTAPIUP=1
+    syslog_netcat "Thrift client API service running on ${FIRST_SEED}. Will check cluster state"
+    check_cluster_state ${FIRST_SEED} 1 1
+    STATUS=$?
+else
+    THRIFTAPIUP=0
+    syslog_netcat "Thrift client API service down on ${FIRST_SEED}. Will unconditionally restart Cassandra"    
+fi
+
+if [[ $STATUS -ne 0 ]]
+then 
+    syslog_netcat "Starting cassandra on ${SHORT_HOSTNAME}" 
+    sudo service cassandra restart 
+
+    if [[ $THRIFTAPIUP -eq 1 ]]
+    then
+        check_cluster_state ${FIRST_SEED} 10 20
+        STATUS=$?
+        if [[  $STATUS -eq 0 ]]
+        then 
+            syslog_netcat "Cassandra cluster fully formed. All nodes registered"
+        else
+            syslog_netcat "Failed to form Cassandra cluster! - NOK"    
+        fi
+    else
+        syslog_netcat "Cassandra service restarted."
+        STATUS=0        
+    fi
+else
+    if [[ $THRIFTAPIUP -eq 1 ]]
+    then        
+        syslog_netcat "Cassandra cluster fully formed. All nodes registered. Bypassing Cassandra restart."
+        STATUS=0
+    fi
+fi
 
 provision_application_stop $START
-exit 0
+exit $STATUS
