+++ scripts/cassandra_ycsb/cb_restart_seed.sh
@@ -24,18 +24,24 @@ START=`provision_application_start`
 
 SHORT_HOSTNAME=$(uname -n| cut -d "." -f 1)
 
-sudo mkdir -p ${CASSANDRA_DATA_DIR}
-
 VOLUME=$(get_attached_volumes)
 if [[ $VOLUME != "NONE" ]]
 then
-    if [[ $(check_filesystem $VOLUME) == "none" ]]
-    then
-        syslog_netcat "Creating $CASSANDRA_DATA_FSTYP filesystem on volume $VOLUME"
-        sudo mkfs.$CASSANDRA_DATA_FSTYP $VOLUME
-    fi
-    syslog_netcat "Making $FSTYP filesystem on volume $VOLUME accessible through the mountpoint ${CASSANDRA_DATA_DIR}"
-    sudo mount $VOLUME ${CASSANDRA_DATA_DIR}
+	if [[ $(sudo mount | grep $VOLUME | grep -c $CASSANDRA_DATA_DIR) -eq 0 ]]
+	then
+		sudo service cassandra stop
+		
+		sudo mkdir -p ${CASSANDRA_DATA_DIR}
+				
+	    if [[ $(check_filesystem $VOLUME) == "none" ]]
+	    then
+	        syslog_netcat "Creating $CASSANDRA_DATA_FSTYP filesystem on volume $VOLUME"
+	        sudo mkfs.$CASSANDRA_DATA_FSTYP $VOLUME
+	    fi
+	    
+	    syslog_netcat "Making $FSTYP filesystem on volume $VOLUME accessible through the mountpoint ${CASSANDRA_DATA_DIR}"
+	    sudo mount $VOLUME ${CASSANDRA_DATA_DIR}
+	fi
 fi
 
 CASSANDRA_REPLICATION_FACTOR=$(get_my_ai_attribute_with_default replication_factor 4)
@@ -100,26 +106,53 @@ syslog_netcat "my ip : $MY_IP"
 #
 # Start the database
 #
-syslog_netcat "Starting cassandra on ${SHORT_HOSTNAME}" 
-sudo service cassandra start 
+FIRST_SEED=$(echo $seed_ips_csv | cut -d ',' -f 1)
 
-# Give all the Java services time to start
-wait_until_port_open ${MY_IP} 9160 20 5
+syslog_netcat "Performing a quick check on ${SHORT_HOSTNAME} in order to decide on Cassandra restart" 
+check_cluster_state ${FIRST_SEED} 1 1
 
 STATUS=$?
+if [[ $STATUS -ne 0 ]]
+then 
+    syslog_netcat "The exit code of \"check_cluster_state ${FIRST_SEED} 1 1\" was $STATUS. Starting cassandra on ${SHORT_HOSTNAME}" 
+    sudo service cassandra restart 
 
-if [[ ${STATUS} -eq 0 ]]
-then
-    syslog_netcat "Cassandra seed server running"
-else
-    syslog_netcat "Cassandra seed server failed to start"
+    # Give all the Java services time to start
+    wait_until_port_open ${MY_IP} 9160 20 5
+
+    STATUS=$?
+
+    if [[ ${STATUS} -eq 0 ]]
+    then
+        syslog_netcat "Cassandra seed server running"
+        check_cluster_state ${FIRST_SEED} 10 20
+        STATUS=$?
+        if [[  $STATUS -eq 0 ]]
+        then 
+            syslog_netcat "Cassandra cluster fully formed. All nodes registered after Cassandra restart"
+        else
+            syslog_netcat "Failed to form Cassandra cluster! - NOK"    
+        fi          
+    else
+        syslog_netcat "Cassandra seed server failed to start"
+        STATUS=1
+    fi
+else                  
+    syslog_netcat "Cassandra cluster fully formed. All nodes registered. Bypassing Cassandra restart."
+    STATUS=0
 fi
-
+    
 #
 # Init database
 #
-cassandra-cli -f create_keyspace.cassandra
 
+if [[ ${MY_IP} == ${FIRST_SEED} ]]
+then
+    if [[ $(cassandra-cli -h ${MY_IP} -f list_keyspace.cassandra | grep Keyspace | grep -c usertable) -eq 0 ]]
+    then
+        cassandra-cli -h ${FIRST_SEED} -f create_keyspace.cassandra
+    fi
+fi
 provision_application_stop $START
 
 exit ${STATUS}
